refactor: Login Terms Page
